debug display part done

+++ Emulator/Framework/gfx.h
@@ -59,7 +59,7 @@ void GFXCloseWindow(void);
 void GFXRectangle(SDL_Rect *rc,int colour);
 void GFXCharacter(int xc,int yc,int character,int size,int colour,int back);
 void GFXString(int xc,int yc,const char *text,int size,int colour,int back);
-void GFXNumber(int xc,int yc,int number,int base,int width,int size,int colour,int back);
+void GFXNumber(int xc,int yc,long number,int base,int width,int size,int colour,int back);
 int  GFXIsKeyPressed(int character);
 int  GFXToASCII(int ch,int applyModifiers);
 int  GFXTimer(void);
